65528: Renaming/Moving according to feedback; style and links added; chevrons fixed
