+++ src/app/community-list-page/community-list-page.component.html
@@ -1,2 +1,4 @@
-<h2>{{ 'communityList.title' | translate }}</h2>
-<ds-community-list></ds-community-list>
+<div class="container">
+    <h2>{{ 'communityList.title' | translate }}</h2>
+    <ds-community-list></ds-community-list>
+</div>
